Very basic framework for features
Allows features to be defined in their own files (solving scope issues),
require'd, and used. All features must accept a client function, after which
point they can bind listeners to the node-irc client as you would normally.

This framework is nice because it is simple, and I think also easy for
beginners to work on (still looks like code from the node-irc docs). It has
some downsides, for example Features can't really be compared to each other or
removed.